Corrected some warnings caught by Visual Studio 2015, added files needed by OpenStage.

+++ Blob/BlobOutputService/m+mBlobOutputInputHandler.cpp
@@ -145,7 +145,7 @@ bool BlobOutputInputHandler::handleInput(const yarp::os::Bottle &     input,
                         
                         if ((0 < numBytes) && asBytes)
                         {
-                            int retVal = send(_outSocket, asBytes, numBytes, 0);
+                            int retVal = send(_outSocket, asBytes, static_cast<int>(numBytes), 0);
                             
                             cerr << "send--> " << retVal << endl; //!!!!
                             if (0 > retVal)
