+++ LeapMotion/LeapMotionInputService/m+mLeapBlobInputListener.cpp
@@ -287,7 +287,7 @@ void LeapBlobInputListener::onFrame(const Leap::Controller & theController)
                     yarp::os::Bottle message;
                     std::string      buffAsString(outBuffer.str());
                     yarp::os::Value  blobValue(const_cast<char *>(buffAsString.c_str()),
-                                               buffAsString.length());
+                                               static_cast<int>(buffAsString.length()));
 
                     message.add(blobValue);
                     if (! _outChannel->write(message))
