+++ Unreal/UnrealOutputService/m+mUnrealOutputLeapInputHandler.cpp
@@ -398,7 +398,8 @@ bool UnrealOutputLeapInputHandler::handleInput(const yarp::os::Bottle &     inpu
                                     outBuffer << "END" << LINE_END_;
                                     std::string outString(outBuffer.str());
                                     int         retVal = send(_outSocket, outString.c_str(),
-                                                              outString.length(), 0);
+                                                              static_cast<int>(outString.length()),
+															  0);
                                     
                                     cerr << "send--> " << retVal << endl; //!!!!
                                     if (0 > retVal)
